requirements.txt
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0

